refactor(#142/*): photographer success code global/response 로 이동
