+++ src/main/java/org/sopt/snappinserver/global/response/code/photographer/PhotographerSuccessCode.java
@@ -1,4 +1,4 @@
-package org.sopt.snappinserver.api.v1.photographer.code;
+package org.sopt.snappinserver.global.response.code.photographer;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
